requirements.txt
Django>=3.2,<3.3
mysqlclient==2.1.0
django-debug-toolbar
django-crispy-forms
colorful
django_social_share
django-allauth
pyopenssl
django-paypal
sentry-sdk
django-crontab
celery
redis
django-celery-beat
django-bootstrap-modal-forms
crispy-bootstrap5
django-storages
boto3
PyGithub
Django-filter
Pillow
stripe
wagtail>=2.15,<2.16
puput==1.1.3
django-crispy-forms
django-bootstrap-modal-forms
crispy-bootstrap5